SakthiDharanMA/requirements.txt
Flask==3.1.0
tensorflow==2.18.0
numpy==1.24.2
pillow==9.2.0
pandas==1.5.3
scikit-learn==1.2.0
matplotlib==3.6.2
seaborn==0.12.0
Flask-WTF==0.16.0
Flask-Login==0.6.0
gunicorn==21.1.0
